// 03 - Canvas 응용/03-01 Particle Network Animation/연습/30-particle-network-canvas/src/components/types.ts
export type ParticleOptions = {
  amount: number;
  color: string;
  defaultRadius: number;
  variantRadius: number;
  defaultSpeed: number;
  variantSpeed: number;
};

export type Position = {
  x: number;
  y: number;
};

export type LineHubOptions = {
  color: string;
  lineRadius: number;
};

export type RGB = {
  r: string;
  g: string;
  b: string;
};